// webapp/src/api/model/itemStack.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { Item } from './item';
import type { Upgrade } from './upgrade';

export interface ItemStack {
  item: Item;
  remainingUpgradeSlots?: number;
  stackSize?: number;
  upgradeSlots?: number;
  upgrades: Upgrade[];
}